Update docs for Traits -> Models/Concerns namespace move
