+++ resources/boost/guidelines/core.blade.php
@@ -18,7 +18,7 @@ Activity logging package for Laravel. Logs model events and manual activities to
 Add to models to automatically log create/update/delete events. Optionally implement `getActivitylogOptions()` to configure which attributes to track (defaults to logging events without attribute changes).
 
 ```php
-use Spatie\Activitylog\Traits\LogsActivity;
+use Spatie\Activitylog\Models\Concerns\LogsActivity;
 use Spatie\Activitylog\Support\LogOptions;
 
 class Article extends Model
