+++ tests/Models/User.php
@@ -4,7 +4,7 @@ namespace Spatie\Activitylog\Test\Models;
 
 use Illuminate\Contracts\Auth\Authenticatable;
 use Illuminate\Database\Eloquent\Model;
-use Spatie\Activitylog\Traits\CausesActivity;
+use Spatie\Activitylog\Models\Concerns\CausesActivity;
 
 class User extends Model implements Authenticatable
 {
